Split registry modifier header in separated headers in concepts/modifiers subfolders
Linked: #31
