+++ src/ecstasy/query/concepts/CMakeLists.txt
@@ -5,5 +5,7 @@ set(SRC
     ${SRC}
     ${INCROOT}/Queryable.hpp
     ${INCROOT}/QueryableNeedAdjust.hpp
+    ${INCROOT}/QueryDataHolder.hpp
+    ${INCROOT}/util.hpp
     PARENT_SCOPE
 )
