+++ src/ecstasy/registry/concepts/CMakeLists.txt
@@ -1,0 +1,9 @@
+set(INCROOT ${INCROOT}/concepts)
+set(SRCROOT ${SRCROOT}/concepts)
+
+set(SRC
+    ${SRC}
+    ${INCROOT}/ComponentType.hpp
+    ${INCROOT}/RegistryModifier.hpp
+    PARENT_SCOPE
+)
