+++ src/ecstasy/registry/modifiers/CMakeLists.txt
@@ -3,8 +3,6 @@ set(SRCROOT ${SRCROOT}/modifiers)
 
 set(SRC
     ${SRC}
-    ${INCROOT}/Modifier.hpp
-    ${INCROOT}/ModifiersList.hpp
     ${INCROOT}/Not.hpp
     PARENT_SCOPE
 )
